requirements.txt
gymnasium[mujoco]>=0.29
mujoco>=3.0.0
stable-baselines3>=2.3.0
hydra-core>=1.3.2
omegaconf>=2.3.0
tensorboard>=2.16
numpy>=1.24
